Add Beamer theme for the US Air Force Academy
This change comprises the initial implementation of a Beamer theme for
the US Air Force Academy. The theme is derived from the theme for
Johns Hopkins University and uses the "bluejay" color theme.

The logo is a Scalable Vector Graphics (SVG) image from the Academy's
website: https://www.usafa.edu/.