Do not increment table counter for records
The longtable package (implicitly being used by ltablex) implicitly
increments the table counter as part of the longtable environment.
This change effectively negates that increment because the scholarship
and service tables are not numbered -- at least not like other tables
in the document.

Stack Exchange: https://tex.stackexchange.com/a/234045