Move code to correct package
It wasn't clear why \bibentry was redefined to be robust (hence the
question "TODO: Must bibentry be robust?") until removing it revealed
that it cannot be used in the argument to \supplement to embed a
citation as the title of the material being included. Consequently,
this change moves the requisite code to the supplement package.