Remove application instructions
This change removes the instructions for applying for academic
promotion. Although it may be helpful to include the instructions with
the promotion package, ensuring that the instructions are up-to-date
would be a difficult task -- it's better to simply point faculty to
the appropriate source for such files.

An alternative to removing the instructions is to move them to the
template. Interspersing them in the context of an example package for
academic promotion would undoubtedly be useful because faculty could
refer to them directly rather than a separate document, but the issue
of maintainability trumps this benefit (at least currently).
Nevertheless, this change also adds a README to the template with the
checklist from the instructions, as this checklist will undoubtedly be
useful on its own.