Remove documentation dependency on packages
This change removes the dependency on class (*.cls) and style (*.sty)
files in the same directory for package documentation. Removing the
dependency prevents recompiling the documentation for all the packages
in a bundle (when there are multiple documented TeX (*.dtx) files in
the same directory) when any one of those package changes.