Do not convert title to PDF string for TOC
Invoking \pdfstringdef (from the hyperref package) to convert the
supplement title to a plain text string for the table of contents
(TOC) has unintended side effects, such as the removal of parentheses.
This change removes that code, which forces the user to specify the
table of contents entry manually when the title is not plain text.