Add examples of supplemental material
This change adds the graded-review package as an example of
supplemental material related to teaching and moves the promotion
package to the supplemental material for service.

In addition, this change (partly) eliminates the dependency on files
outside this directory by using a symbolic link to each of the
aforementioned package's documentation.