Revise promotion template supplementary material (#34)
The promotion template provides a comprehensive example of how to use
the packages in the promotion bundle, but examples of supplementary
material were limited. This change adds the graded-review package as
an example of supplemental material related to teaching and moves the
packages in the promotion bundle to the supplemental material for
service.

In addition, this change (mostly) eliminates the dependency on files
outside this directory by using a symbolic link to each of the
aforementioned package's documentation. Thus, the template can be used
with limited modifications provided that the symbolic link is replaced
with the actual document (e.g., using `cp --dereference`).