Prevent infinite recursion for promotion template
The promotion template includes the default Makefile (`Makefile.mk`)
to avoid repeating all the rules to compile LaTeX documents.
Unfortunately, the search path includes the names of packages, which
requires the invocation of `make list`, and executing this command
effectively results in infinite recursion. Hence, this change guards
it by not searching for packages if `Makefile.mk` is in the same
directory.