Allow TEXINPUTS to shadow packages in texmf/
This change modifies the TEXINPUTS environment variable so that can be
overridden to alter the order in which packages are found. More
specifically, it enables a package in another directory (e.g., ../) to
be used instead of a package with the same name in the texmf directory
at the root of the repository.