Create template for end slide in presentation
This change comprises a Beamer template ("end page" that is loosely
analogous to the existing "title page" template) for the last slide of
a presentation and displays the United States Air Force Academy
shield. This template may be used as follows:

    \usebeamertemplate{end page}

The slide is automatically inserted at the end of presentations that
use the usafa.edu presentation theme.