Move shared LaTeX files to texmf/include/
This change moves commonly-used LaTeX files, such as acronym
definitions and the BibTeX database, to the texmf/include directory.
Doing so has two benefits. First, it avoids cluttering the top-level
directory with these files. Second, it eliminates the need to provide
the (relative) path to these files, which minimizes the need for
updates when a document is moved to a different directory.

The choice of the destination directory (i.e., texmf/include/) loosely
follows the guidance of the TeX Directory Structure (TDS) and the
Filesystem Hierarchy Standard. The texmf directory provides a
repository-specific TDS tree for packages that aren't available from
the Comprehensive TeX Archive Network (CTAN). The subdirectory,
include, is used for standard include files according to the
Filesystem Hierarchy Standard, and although this directory is not
specified as a standard top-level TDS directory, the TDS documentation
explicitly allows that "additional top-level directories may be
present."