Replace seal with logo in Beamer theme (#60)
According to the United States Air Force Academy brand guidelines, the
official seal "is used for official documents and communications" --
i.e., official or ceremonial documents. Moreover, permission must be
granted to use the official emblem and coat of arms.

This change replaces the seal at the conclusion of the presentation
with the official logo.